pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "easy-sdk"
version = "0.1.0"
description = "AI-powered Django API documentation generator with Sphinx and TypeScript support"
readme = "README.md"
license = {text = "MIT"}
authors = [
    {name = "Django Docs Generator Team", email = "contact@example.com"}
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Framework :: Django",
    "Framework :: Django :: 3.2",
    "Framework :: Django :: 4.0",
    "Framework :: Django :: 4.1",
    "Framework :: Django :: 4.2",
    "Framework :: Django :: 5.0",
    "Topic :: Documentation",
    "Topic :: Software Development :: Documentation",
]
keywords = ["django", "documentation", "api", "sphinx", "typescript", "ai", "drf"]
requires-python = ">=3.8"
dependencies = [
    "Django>=3.2",
    "djangorestframework>=3.12",
    "Sphinx>=6.2.1",
    "alabaster>=0.7.13",
    "Jinja2>=3.1.0",
    "click>=8.0.0",
    "pydantic>=2.0.0",
    "astroid>=2.15.0",
    "rich>=13.0.0",
    "typing-extensions>=4.0.0",
]

[project.optional-dependencies]
ai = [
    "openai>=1.0.0",
    "anthropic>=0.7.0",
]
dev = [
    "pytest>=7.0.0",
    "pytest-django>=4.5.0",
    "pytest-cov>=4.0.0",
    "black>=23.0.0",
    "isort>=5.12.0",
    "flake8>=6.0.0",
    "mypy>=1.0.0",
    "pre-commit>=3.0.0",
    "openai>=1.0.0",
    "anthropic>=0.7.0",
]
test = [
    "pytest>=7.0.0",
    "pytest-django>=4.5.0",
    "pytest-cov>=4.0.0",
    "tox>=4.0.0",
]
docs = [
    "sphinx-rtd-theme>=1.2.0",
    "myst-parser>=2.0.0",
]

[project.urls]
Homepage = "https://github.com/easy-sdk/django-api-docs-generator"
Documentation = "https://django-api-docs-generator.readthedocs.io"
Repository = "https://github.com/easy-sdk/django-api-docs-generator"
Issues = "https://github.com/easy-sdk/django-api-docs-generator/issues"

[project.scripts]
easy-sdk = "easy_sdk.cli:main"

[tool.setuptools]
packages = ["easy_sdk"]
package-dir = {"" = "src"}

[tool.setuptools.package-data]
easy_sdk = ["templates/**/*", "static/**/*"]

[tool.black]
line-length = 88
target-version = ['py38']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''

[tool.isort]
profile = "black"
multi_line_output = 3
line_length = 88

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py", "*_test.py"]
addopts = "--cov=src/easy_sdk --cov-report=html --cov-report=term-missing"
DJANGO_SETTINGS_MODULE = "tests.django_settings"